Directory refactor
