+++ internal/repository/repositories/sign_profile_service.go
@@ -1,4 +1,4 @@
-package services
+package repositories
 
 import (
 	"context"
