+++ internal/repository/repositories/sign_request_service_test.go
@@ -1,4 +1,4 @@
-package services
+package repositories
 
 import (
 	"os"
